Add files via upload
I really enjoyed the online course. The work load was just enough, so i could finish everything with enough time, learn about the topics and not feel over loaded and rushed. Overall I really liked the class because all the lectures and tests were straight forward. I really enjoyed the class and the format it was presented in. For me, I learn and retain much more through an online class due to the fact you can do the course as an open book. It is well known that Excel is a software program created by Microsoft that uses spreadsheets to organize numbers and data with formulas and functions. Excel analysis is ubiquitous around the world and used by businesses of all sizes to perform financial analysis.